requirements.txt
fastapi==0.103.0
uvicorn==0.22.0
pydantic==2.3.0
python-frontmatter==1.0.0
python-slugify==8.0.1
mangum
boto3
